+++ buildSrc/src/main/kotlin/io/github/androidpoet/quickactions/Configuration.kt
@@ -2,7 +2,7 @@ package io.github.androidpoet.quickactions
 
 object Configuration {
     const val GROUP = "io.github.androidpoet"
-    const val VERSION = "0.1.0"
+    const val VERSION = "0.2.0"
     const val COMPILE_SDK = 36
     const val MIN_SDK = 26
 }