+++ sample/composeApp/iosApp/iosApp/iOSApp.swift
@@ -3,9 +3,6 @@ import SwiftUI
 
 @main
 struct iOSApp: App {
-    // Wires quick-action delivery (scene connect + performActionFor) to QuickActions.
-    @UIApplicationDelegateAdaptor(QuickActionsAppDelegate.self) var delegate
-
     var body: some Scene {
         WindowGroup {
             ComposeView().ignoresSafeArea()