+++ sample/composeApp/src/iosMain/kotlin/io/github/androidpoet/quickactions/sample/Platform.ios.kt
@@ -1,5 +1,8 @@
 package io.github.androidpoet.quickactions.sample
 
+import io.github.androidpoet.quickactions.QuickActions
 import platform.UIKit.UIDevice
 
 actual fun platformName(): String = "${UIDevice.currentDevice.systemName} ${UIDevice.currentDevice.systemVersion}"
+
+actual fun deliveryInstalled(): Boolean = QuickActions.isDeliveryInstalled